backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Pillow==11.2.1
djoser==2.3.1
psycopg2==2.9.10
gunicorn==20.1.0
Flake8==7.2.0
drf-extra-fields==3.7.0
django-filter==25.1
